% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeScale.R
\name{markerSizeScale}
\alias{markerSizeScale}
\title{markerSizeScale}
\usage{
markerSizeScale(size, maxScale = 20)
}
\arguments{
\item{size}{size vector}

\item{maxScale}{max size of marker (in px)}
}
\value{
Returns a vector which has been scaled between 5 and maxScale
}
\description{
Internal function to scale marker sizes
}

