% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplotly.R
\name{scatterplotly}
\alias{scatterplotly}
\title{scatterplotly}
\usage{
scatterplotly(x, y, xlab = NULL, ylab = NULL, theme, groupcolor, groupsize,
  groupsymbol, title = "", subtitle = "", title.size = 30,
  title.fontface = "Times New Roman", description = "", desc.x = 0.7,
  desc.y = 0.2, desc.size = 12, desc.fontface = "arial", markersize = 5,
  markeropacity = 0.8, markertype = "circle", markerlinewidth = 1,
  markerlinecolor = "black", width = NULL, height = NULL,
  xaxis.range = NULL, yaxis.range = NULL, legendorientation = "v", ...)
}
\arguments{
\item{x}{A vector of x-coordinates}

\item{y}{A vector of y-coordinates}

\item{xlab}{x-axis label (optional)}

\item{ylab}{y-axis label (optional)}

\item{theme}{color theme (see details below)}

\item{groupcolor}{a vector specifying group level color mapping}

\item{groupsize}{a vector specifying group level size mapping}

\item{groupsymbol}{a vector specifying group level marker-type mapping}

\item{title}{Chart title}

\item{subtitle}{Chart subtitle (optional)}

\item{title.size}{Chart title size (default: 30)}

\item{title.fontface}{Chart title (and subtitle) font family(default: Times New Roman)}

\item{description}{Character vector to specify additional description (optional)}

\item{desc.x}{Description annotation's x position (default: 0.7)}

\item{desc.y}{Description annotation's y position (default: 0.2)}

\item{desc.size}{Description annotation's font size (default: 12)}

\item{desc.fontface}{Description annotation's font family (default: arial)}

\item{markersize}{Point size (in px) (default: 5)}

\item{markeropacity}{Point alpha (0  1) (default: 1)}

\item{markertype}{Point style (default: circle)}

\item{markerlinewidth}{Width of line around point (default: 1)}

\item{markerlinecolor}{Color of line around point (default: Black)}

\item{width}{Canvas width (default: 800)}

\item{height}{Canvas height (default: 600)}

\item{xaxis.range}{xaxis limits (a numeric vector of size two specifying lower and upper limits)}

\item{yaxis.range}{yaxis limits (a numeric vector of size two specifying lower and upper limits)}

\item{legendorientation}{specify the location of the legend ("v" or "h)}

\item{...}{Other arguments that can be passed to plot_ly()}
}
\value{
Returns a plotly built object. Use $ for further manipulations and edits.
}
\description{
Create scatterplots using \href{https://plot.ly/r}{plotly's R API}.
Type \code{scatterplotly()} for visual help.
}
\details{
Internally, \code{plot_ly()} is used to generate all plots. All of the above arguments are
mapped to specific arguments to \code{plot_ly()}. The function assigns default values to
many \code{plot_ly()} arguments like marker-size, marker-opacity etc. to allow for
an easier interface.

Type \code{plotlyvis::scatterplotly} to see the source code. Alternatively you can press F2
if using R-Studio.

Currently, four color themes are built in: BLUE, YELLOW, LIGHTGRAY and DARKGRAY.
}
\note{
The \emph{title}, \emph{subtitle} and \emph{description} are added as annotations.
By default the \code{xref} and \code{yref} arguments are set to \emph{paper}.
See \href{https://plot.ly/r/reference/#layout-annotationsplot}{plotly - annotations}
for more details.
}
\examples{
library(plotlyvis)

# Simple scatter plot
scatterplotly(x = rnorm(1000), y = rnorm(1000), title = "Simple scatter plot")

# Group level mappings
n = 500
x = seq(-1, 1, length.out = n)
y = round(2 * x^2 + rnorm(n, sd = 0.25),2)
size = c(1:250, 250:1)
color = sample(LETTERS[1:4], size = n, replace = TRUE)
symbol = sample(letters[1:3], size = n, replace = TRUE)
scatterplotly(x = x, y = y,
              groupcolor = color,
              groupsize = size,
              groupsymbol = symbol,
              title = "Advanced scatterplot",
              subtitle = "Scatterplot with group level color, size and symbol mappings")

#Theme
scatterplotly(x = rnorm(1000), y = rnorm(1000), title = "Simple scatter plot", theme = "BLUE")

}
\author{
Riddhiman Roy
}
\seealso{
\code{\link[plotly]{plot_ly}}
\code{\link[plotly]{plotly_build}}
}

