% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WorldEducationIndicators}
\alias{WorldEducationIndicators}
\title{Education statistics}
\format{A data frame with 3630 rows and 6 variables:
\itemize{
\item Country: Country of observation
\item Year: Year (2010 - 2014)
\item Adult.literacy.rate: Percentage of population age 15 and above who can read and write
divided by the corresponding age group population multiplied by 100.
\item GDP.at.market.prices: GDP in constant 2005 U.S. dollars.
\item Government.education.expenditure: Total general (local, regional and central)
government expenditure on education expressed as a percentage of total general government
expenditure.
\item Youth.literacy.rate: Number of people age 15 to 24 years who can both read and write
divided by the total population in the same age group multiplied by 100.
}}
\source{
\url{http://data.worldbank.org/data-catalog/ed-stats}
}
\usage{
WorldEducationIndicators
}
\description{
A dataset containing statistics related to literacy rates for various countries
across different years.
}
\keyword{datasets}

